Added bindArrayValue which binds the parameters according to their types.
$stmt->execute($param); would bind all parameters as PDO::PARAM_STR, which is incorrect for boolean FALSE value. 
If FALSE is being inserted to the database as PDO::PARAM_STR, empty string is inserted, and not 0 as it is supposed to be.

I decided to evaluate the null parameters as PDO::PARAM_STR and not as PDO::PARAM_NULL, because of problems with mysql null column value recognition.

Prepared statements with zero-indexed array of parameters and associative arrays are kept in mind.

Related forum post: http://doophp.com/forum/viewtopic.php?f=10&t=1693